requirements-dev.txt
-r requirements.txt

ipdb 
autoflake
black
flake8
isort
pytest
pytest-cov 
pytest-flask 
scrapy-mock